Make ServerCallInfo abstract
